--- src/Components/GenerateTornPaperPiece/Paper.module.css
@@ -1,23 +1,0 @@
-
-
-
-.paper{
-  background: #;
-  max-width: 20%;
-  padding: 20px;
-  position: relative;
-  width: min-content;
-  box-shadow: 0 3px 5px rgba(0,0,0,0.05);
-}
-  
-.paper::after {
-  content: "";
-  height: 20px;
-  display: block;
-  background: url(https://s3-us-west-2.amazonaws.com/s.cdpn.io/3/rip.svg) bottom left;
-  /* background-size: 100%; */
-  position: absolute;
-  top: 100%;
-  left: 0;
-  width: 100%;
-}
