+++ src/Сomponents/ButtonCutUp/ButtonCutUp.js
@@ -24,5 +24,3 @@ export class ButtonCutUp extends React.Component {
     );
   }
 }
-
-
