+++ src/Сomponents/ButtonCutUp/cutRandomPieces.js
