+++ src/Сomponents/ButtonCutUp/index.js
