+++ src/Сomponents/Paper/GenerateTornPaperPiece.js
