+++ src/Сomponents/Paper/Paper.js
