+++ src/Сomponents/Paper/Paper.module.css
