+++ src/Сomponents/Paper/index.js
