+++ src/Сomponents/Text/text.module.css
